% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aic.fitames.R
\name{aic.fitames}
\alias{aic.fitames}
\title{Akaike measures for fitames objects}
\usage{
aic.fitames(..., model.names, C = T, wis.plot = F, emql.all = F,
  SEE = 2)
}
\arguments{
\item{...}{One or more "fitames" class objects,
usually fitted obtained by fitting models using \code{\link{fit.fitames}}.}

\item{model.names}{Character (optional), specifying names for each model.}

\item{C}{Logical. If FALSE, computes AIC or QAIC
depening on the fitted model. If TRUE (default) computes the AICc or QAICc.}

\item{wis.plot}{Logical. If TRUE, creates an index plot using the
Akaike weights, and highlights those satisfying Delta < SEE. Defaults to FALSE.}

\item{emql.all}{Logical. If TRUE, computes the AIC or AICc based on the
extended quasi-likelihood function for each model. Defaults to FALSE.}

\item{SEE}{Numeric. Positive value indicating the strength of empyrical evidence (SEE).
Default value is 2.}
}
\description{
Extracts the Akaike information criterion (AIC) of fitted models,
its second-order, bias-corrected version for small samples (AICc),
or extracts the QAIC/QAICc, which are based on the quasi-likelihood approach.
If several models are supplied, then Delta values are also
computed and plots the Akaike weights.
}
\examples{
aic.fitames()
}
\author{
Davi Butturi-Gomes

Silvio S. Zocchi
}
\seealso{
\code{\link{fit.ames}}
}
\keyword{aic}
\keyword{fitames}

