% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ames-deprecated.R, R/dgenpois.R, R/rgenpois.R
\docType{package}
\name{ames-deprecated}
\alias{ames-deprecated}
\alias{ames-deprecated-package}
\alias{dgenpois}
\alias{rgenpois}
\title{Deprecated function(s) in the ames package}
\usage{
dgenpois(x, mu, phi)

rgenpois(n, mu, phi)
}
\description{
These functions are provided for compatibility with older version of
the ames package.  They may eventually be completely
removed.
}
\section{Details}{

\tabular{rl}{
  \code{\link{dgp1}} \tab now a synonym for \code{\link{dgenpois}}\cr
  \code{\link{rgp1}} \tab now a synonym for \code{\link{rgenpois}}\cr
}
}

