% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/control.fitames.R
\name{control.fitames}
\alias{control.fitames}
\title{Controls of fitting procedures in fit.ames}
\usage{
control.fitames(convergence.tol, max.it, zerocounts, strain, cfu, alpha, delta,
  allow.jitter, max.jitter, dec, tau.profile, conv.crit, halfstepping, stepsize,
  diagW.min, detas.method, nsim)
}
\arguments{
\item{...}{}
}
\description{
Controls of fitting procedures in fit.ames
}
\examples{
fit.ames()
}
\author{
Davi Butturi-Gomes

Silvio S. Zocchi
}

