% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dgenpois.R
\name{dgenpois}
\alias{dgenpois}
\title{Type-I generalized Poisson distribution}
\usage{
dgenpois(x, mu, phi)
}
\arguments{
\item{...}{}
}
\description{
Type-I generalized Poisson distribution
}
\examples{
fit.ames()
}
\author{
Davi Butturi-Gomes

Silvio S. Zocchi
}

