% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diagnostics.fitames.R
\name{diagnostics.fitames}
\alias{diagnostics.fitames}
\title{Model diagnostics for fitames objects}
\usage{
diagnostics.fitames(obj, lmaxes.plot = T, curvs.plot = T,
  nuisance.plot = T, eta.plot = NULL)
}
\arguments{
\item{...}{}
}
\description{
Model diagnostics for fitames objects
}
\examples{
iv.ames()
}
\author{
Davi Butturi-Gomes

Silvio S. Zocchi
}
\seealso{
\code{\link{plot.fitames}}
}

