% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dskellam2.R
\name{dskellam2}
\alias{dskellam2}
\title{Skellam distribution}
\usage{
dskellam2(x, mu1, mu2, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{mu1}{First parameter for the mean of the distribution}

\item{mu2}{Second parameter for the mean of the distribution}

\item{log}{Logical. If TRUE, returns the density in the natural logarithm scale.
Defaults to FALSE.}
}
\description{
Density of the Skellam distribution, slightly modified from code{dskellam} \{VGAM\}.
}
\author{
Davi Butturi-Gomes

Silvio S. Zocchi
}
\seealso{
\code{\link[VGAM]{dskellam}}
}

