% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit.ames.R
\name{fit.ames}
\alias{fit.ames}
\title{Fitting models for Ames test data}
\usage{
fit.ames(data, count, dose, lh = c("poisson", "nb2", "quasipoisson",
  "quasipower", "skellam", "nlvnb2"), predictor = c("krewski", "myers",
  "bernstein", "breslow", "stead", "margolin"), link = log, form1, form2,
  theta.start = "smart", dispersion.start = 1, dispersion.method = c("ml",
  "mpl", "profile"), controls = control.fitames(), quiet = F, ...)
}
\arguments{
\item{data}{a matrix or a data frame. The standard use of this structure inside the function is
revernt count in the first column and doses placed in the second column. If not supplied,
the function will searc for the arguments \code{count} and \code{dose}.}

\item{count}{Integer. A vector of length equal to the number of observations, describing the
number of revertent colonies per plate. If both \code{data} and \code{count} are supplied, then
\code{count} is ignored.}

\item{dose}{Numeric. A vector of same length as \code{count}, describind the doses in each plate.}

\item{lh}{Character. A sequence of characters describing which random component should used
for model fitting. For GNLMs, the valid sequences are "poisson", "nb2", "quasipoisson" and
"quasipower". For VGNLMs, the valid sequences are "skellam" and "nlvnb2". See \bold{Details}.}
}
\description{
This is the main function of the "ames" package -- potentially, fits several different models
to the data. Its output is an object of class "fitames", used by most of the other functions
available in the package.
}
\details{
The families described by the argument \code{lh} differ either by the complete
especification of likelihood function, or by the variance function formula. If "poisson"
is chosen, then a complete Poisson likelihood is used, whth \emph{phi} fixed (=1) and
variance function \emph{V(mu)=mu}. If "nb2", the assumptions depend on the argument
\code{dispersion.method}. Thus, for "ml", then a complete negative binomial type-II
likelihood is used, \emph{phi} is fixed (=1), and \emph{V(mu)=mu(1+mu/k)}; for "mpl", instead,
\emph{phi=mu(1+mu/k)} and \emph{V(mu)=1} and the estimation of \emph{k} is obtained by
maximum pseudo-likelihood. If "quasipoisson", no likelihood function is completely especified,
and the estimation procceeds using the maximum quasi-likelihood with \emph{V(mu)=mu},
\emph{Var(Y)=phi*V(mu)} and \emph{phi} an unknown constant, estimated by a moments method
(generalized Pearson statistic). If "quasipower".
}
\examples{
fit.ames()
}
\author{
Davi Butturi-Gomes

Silvio S. Zocchi
}

