% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iv.ames.R
\name{iv.ames}
\alias{iv.ames}
\title{Compute inital values for fitting models to Ames test data}
\usage{
iv.ames(data, count, dose, predictor = c("krewski", "myers", "margolin",
  "stead"), link = log, controls = control.fitames(), interval = c(-5, 5),
  submod = F, ...)
}
\arguments{
\item{...}{}
}
\description{
Compute inital values for fitting models to Ames test data
}
\examples{
iv.ames()
}
\author{
Davi Butturi-Gomes

Silvio S. Zocchi
}

