% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logLik.fitames.R
\name{logLik.fitames}
\alias{logLik.fitames}
\title{Log-likelihood on convergence of a fitames object}
\usage{
logLik.fitames(obj)
}
\arguments{
\item{...}{}
}
\description{
Log-likelihood on convergence of a fitames object
}
\examples{
fit.ames()
}
\author{
Davi Butturi-Gomes

Silvio S. Zocchi
}

