% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.fitames.R
\name{plot.fitames}
\alias{plot.fitames}
\title{Residual plots and model diagnostics and for fitames objects}
\usage{
\method{plot}{fitames}(obj, smooth = T, deviance.plots = F,
  std.deviance.plots = F, pearson.plots = F, std.pearson.plots = F,
  qqplot = F, lag.plot = F, lag = 1, diag.measures = F, envelope = F,
  nsim = 99, probs = c(0.005, 0.995), all = F, quiet = T)
}
\arguments{
\item{...}{}
}
\description{
Residual plots and model diagnostics and for fitames objects
}
\examples{
iv.ames()
}
\author{
Davi Butturi-Gomes

Silvio S. Zocchi
}

