% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rames.R
\name{rames}
\alias{rames}
\title{Ames test data random generator}
\usage{
rames(dose, theta, dispersion = 1, mu.only = F, lh = c("poisson", "nb2",
  "gp1"), predictor = c("krewski", "myers", "bernstein", "breslow", "stead",
  "margolin", "svetliza"), link = log, controls = control.fitames(),
  quiet = T)
}
\arguments{
\item{...}{}
}
\description{
Ames test data random generator
}
\examples{
iv.ames()
}
\author{
Davi Butturi-Gomes

Silvio S. Zocchi
}

