% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gp1.R
\name{dgp1}
\alias{dgp1}
\alias{rgp1}
\title{Generalized Poisson Type-1 distribution (GP-1)}
\usage{
dgp1(x, mu, phi, log = F)

rgp1(n, mu, phi)
}
\arguments{
\item{x}{Integer, vector of integers or numeric that can be coerced into integers.}

\item{mu}{Positive numeric. Mean parameter of the distribution}

\item{phi}{Positive numeric. Dispersion parameter.}

\item{log}{Logical. If TRUE, return the probabilities in log scale. The default is FALSE.}

\item{n}{Integer or numeric that can be coerced into integer. Number of random deviations
to be generated}
}
\description{
\code{dgp1} - Generalized Poisson type-1 distribution probability mass function

\code{rgp1} - Generalized Poisson type-1 distribution random number generator
}
\examples{
x <- rgp1(1000,mu=mu<-20,phi=phi<-2)
mu; mu*phi # Nominal mean and variance
mean(x);var(x) # "Observed" mean and variance
#
dx <- (dgp1(15:25,mu=mu,phi=phi))
plot(15:25,dx,type='h')
}
\author{
Davi Butturi-Gomes

Silvio S. Zocchi
}
\seealso{
\code{\link{dgp2}}, \code{\link{rgp2}}, \code{\link{dgpg}}, \code{\link{rgpg}}
}

