% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gp2.R
\name{dgp2}
\alias{dgp2}
\alias{rgp2}
\title{Generalized Poisson Type-2 distribution (GP-2)}
\usage{
dgp2(x, mu, phi, log = F)

rgp2(n, mu, phi)
}
\arguments{
\item{x}{Integer, vector of integers or numeric that can be coerced into integers.}

\item{mu}{Positive numeric. Mean parameter of the distribution}

\item{phi}{Positive numeric. Dispersion parameter.}

\item{log}{Logical. If TRUE, return the probabilities in log scale. The default is FALSE.}

\item{n}{Integer or numeric that can be coerced into integer. Number of random deviations
to be generated}
}
\description{
\code{dgp2} - Generalized Poisson type-2 distribution probability mass function

\code{rgp2} - Generalized Poisson type-2 distribution random number generator
}
\examples{
x <- rgp2(1000,mu=mu<-20,phi=phi<-0.01)
mu; mu*((1+phi*mu)^2) # nominal mean and variance
mean(x);var(x) # "observed" mean and variance
#
dx <- (dgp2(10:30,mu=mu,phi=phi))
plot(10:30,dx,type='h')
}
\author{
Davi Butturi-Gomes

Silvio S. Zocchi
}
\seealso{
\code{\link{dgp1}}, \code{\link{rgp1}}, \code{\link{dgpg}}, \code{\link{rgpg}}
}

