% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/skellam2.R
\name{skellam2}
\alias{skellam2}
\title{Skellam Distribution Family Function}
\usage{
skellam2(lmu1 = "loge", lmu2 = "loge", imu1 = NULL, imu2 = NULL,
  nsimEIM = 100, parallel = FALSE, zero = NULL, ...)
}
\description{
Estimates the two parameters of a Skellam distribution by maximum likelihood estimation.
This is to be used along with \code{vglm} function, available within \{VGAM\} package.
}
\author{
Davi Butturi-Gomes

Silvio S. Zocchi
}
\seealso{
\code{\link[VGAM]{skellam}}, \code{\link[VGAM]{vglm}}
}

