% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mmd.fitames.R
\name{mmd.fitames}
\alias{mmd.fitames}
\title{Maximum Mutagenic Dose (MMD)}
\usage{
mmd.fitames(..., model.names, akaike.weights = F, emql.all = F)
}
\arguments{
\item{...}{One or more "fitames" class objects, usually fitted obtained by fitting models using fit.ames().}

\item{model.names}{Character (optional), specifying names for each model.}

\item{akaike.weights}{Logical. If TRUE, internally calls aic.fitames(), recovering the Akaike weights and uses these values to compute the multimodel MMD. Defaults to FALSE.}

\item{emql.all}{Logical. Argument passed to aic.fitames(), if akaike.weights is TRUE, otherwise ignored. Defaults to FALSE.}
}
\description{
Computes the MMD values, given one or more previously fitted models
}
\examples{
mmd.fitames()
}
\keyword{ames,}
\keyword{mi}
\keyword{mmd,}

